+++ src/components/posts/PetLost.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-    import ItemFound from "./Item";
+    import ItemFound from "./PostItem";
     import { mapActions, mapState, mapGetters } from "vuex";
     import InfiniteList from "../common/InfiniteList";
     import listMixin from "./mixins/list";
